Refactor by moving functions from app.py to modules
